+++ simaple/simulate/component/__init__.py
@@ -1,9 +1,3 @@
 # ruff: noqa: F401
 
-import simaple.simulate.component.complex_skill
-import simaple.simulate.component.hexa_skill
-import simaple.simulate.component.keydown_skill
-import simaple.simulate.component.mob
-import simaple.simulate.component.periodic_with_finish_skill
-import simaple.simulate.component.skill
-import simaple.simulate.component.triggable_buff_skill
+import simaple.simulate.component.common
