+++ simaple/simulate/component/specific/mechanic.py
@@ -8,9 +8,14 @@ from simaple.simulate.component.base import (
     reducer_method,
     view_method,
 )
-from simaple.simulate.component.entity import Cycle
-from simaple.simulate.component.keydown_skill import Keydown
-from simaple.simulate.component.skill import Cooldown, Lasting, Periodic, SkillComponent
+from simaple.simulate.component.entity import (
+    Cooldown,
+    Cycle,
+    Keydown,
+    Lasting,
+    Periodic,
+)
+from simaple.simulate.component.skill import SkillComponent
 from simaple.simulate.component.trait.impl import (
     BuffTrait,
     CooldownValidityTrait,
