+++ tests/simulate/component/test_periodic_entity.py
@@ -1,6 +1,6 @@
 import pytest
 
-from simaple.simulate.component.skill import Periodic
+from simaple.simulate.component.entity import Periodic
 
 
 @pytest.mark.parametrize(
