+++ tests/simulate/test_base.py
@@ -1,7 +1,7 @@
 from simaple.core.base import ActionStat
 from simaple.simulate.base import AddressedStore, ConcreteStore
 from simaple.simulate.builder import EngineBuilder
-from simaple.simulate.component.skill import AttackSkillComponent
+from simaple.simulate.component.common.attack_skill import AttackSkillComponent
 from simaple.simulate.global_property import GlobalProperty
